pyproject.toml
[tool.poetry]
name = "desafio-backend"
version = "0.1.0"
description = ""
authors = ["PontoTel"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
SQLAlchemy = "^2.0.3"
gitpython = "^3.1.31"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
alembic = "^1.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
